requirements.txt
GitPython==2.1.11
Pillow==4.0.0
#5.2.0
googledrivedownloader==0.3
pathos==0.2.2.1
pandas==0.23.4
opencv-python==3.4.3.18
keras==2.2.4
tensorflow-gpu==1.12.0
